Fix PostGis unit tests
